Project functions encapsulated in a ProjectService class namespace.
