+++ src/gluon/team/MembersAddedToTeam.ts
@@ -10,7 +10,7 @@ import * as _ from "lodash";
 import {OCCommandResult} from "../../openshift/base/OCCommandResult";
 import {BitbucketConfiguration} from "../bitbucket/BitbucketConfiguration";
 import {getProjectDisplayName} from "../project/Project";
-import {gluonProjectsWhichBelongToGluonTeam} from "../project/Projects";
+import {ProjectService} from "../project/ProjectService";
 import {
     ChannelMessageClient,
     handleQMError,
@@ -49,6 +49,9 @@ subscription MembersAddedToTeamEvent {
 `)
 export class MembersAddedToTeam implements HandleEvent<any> {
 
+    constructor(private projectService = new ProjectService()) {
+    }
+
     public async handle(event: EventFired<any>, ctx: HandlerContext): Promise<HandlerResult> {
         logger.info(`Ingested MembersAddedToTeamEvent event: ${JSON.stringify(event.data)}`);
 
@@ -59,10 +62,10 @@ export class MembersAddedToTeam implements HandleEvent<any> {
 
             let projects;
             try {
-                projects = await gluonProjectsWhichBelongToGluonTeam(ctx, team.name);
+                projects = await this.projectService.gluonProjectsWhichBelongToGluonTeam(ctx, team.name);
             } catch (error) {
                 // TODO: We probably dont want to have the gluonProjectsWhichBelong to team thing catch these errors for events
-                return logErrorAndReturnSuccess(gluonProjectsWhichBelongToGluonTeam.name, error);
+                return logErrorAndReturnSuccess(this.projectService.gluonProjectsWhichBelongToGluonTeam.name, error);
             }
 
             const bitbucketConfiguration = this.getBitbucketConfiguration(membersAddedToTeamEvent);
